Separate simulation loadtest and tutorial loadtest.
